Manipulate subset content

+++ servermodule/panoptesserver/subset_store.py
@@ -23,7 +23,7 @@ def ResponseExecute(returndata, calculationObject):
     subsetid = DQXDbTools.ToSafeIdentifier(returndata['subsetid'])
     method = DQXDbTools.ToSafeIdentifier(returndata['method'])
 
-    if method not in ['add', 'replace']:
+    if method not in ['add', 'replace', 'remove']:
         raise Exception('Invalid store selection method')
 
     calculationObject.Log('==== storing subset')
@@ -35,8 +35,10 @@ def ResponseExecute(returndata, calculationObject):
 
     if not(isNumericalKey):
         valuesString = '("{0}", {1})'
+        deleteString = '"{0}"'
     else:
         valuesString = '({0}, {1})'
+        deleteString = '{0}'
     subsetTable = tableid + '_subsets'
 
 
@@ -47,37 +49,60 @@ def ResponseExecute(returndata, calculationObject):
         # cur.execute(sqlstring)
         # cur.commit()
 
-        if method == 'replace':
-            cur.execute('DELETE FROM {0} WHERE subsetid={1}'.format(DBTBESC(subsetTable), subsetid))
-            cur.commit()
 
-        existingRecordMap = {}
-        if method == 'add':
-            calculationObject.Log('Determining current subset content...')
-            cur.execute('SELECT {0} FROM {1} WHERE subsetid={2}'.format(DBCOLESC(keyid), DBTBESC(subsetTable), subsetid))
-            for row in cur.fetchall():
-                existingRecordMap[row[0]] = True
-
-        if len(datastring) > 0:
-            keys = datastring.split('\t')
-            def submitkeys(keylist):
-                if len(keylist) > 0:
-                    sqlstring = "INSERT INTO {0} VALUES ".format(DBTBESC(subsetTable))
-                    sqlstring += ', '.join([valuesString.format(key, subsetid) for key in keylist])
-#                    print(sqlstring)
-                    cur.execute(sqlstring)
-                    cur.commit()
-            keysublist = []
-            keyNr = 0
-            for key in keys:
-                if key not in existingRecordMap:
+        if method == 'remove':
+            if len(datastring) > 0:
+                keys = datastring.split('\t')
+                def delkeys(keylist):
+                    if len(keylist) > 0:
+                        keystr = ', '.join(['"'+deleteString.format(key)+'"' for key in keylist])
+                        sqlstring = "DELETE FROM {0} WHERE (subsetid={1}) AND ({2} IN ({3}))".format(DBTBESC(subsetTable), subsetid, DBCOLESC(keyid), keystr)
+                        cur.execute(sqlstring)
+                        cur.commit()
+                keysublist = []
+                keyNr = 0
+                for key in keys:
                     keysublist.append(key)
-                    if len(keysublist) >= 500:
-                        submitkeys(keysublist)
+                    if len(keysublist) >= 200:
+                        delkeys(keysublist)
                         keysublist = []
-                        calculationObject.SetInfo('Storing', keyNr*1.0/len(keys))
+                        calculationObject.SetInfo('Removing', keyNr*1.0/len(keys))
                     keyNr += 1
-            submitkeys(keysublist)
+                delkeys(keysublist)
+
+
+        if (method == 'replace') or (method == 'add'):
+
+            if method == 'replace':
+                cur.execute('DELETE FROM {0} WHERE subsetid={1}'.format(DBTBESC(subsetTable), subsetid))
+                cur.commit()
+
+            existingRecordMap = {}
+            if method == 'add':
+                calculationObject.Log('Determining current subset content...')
+                cur.execute('SELECT {0} FROM {1} WHERE subsetid={2}'.format(DBCOLESC(keyid), DBTBESC(subsetTable), subsetid))
+                for row in cur.fetchall():
+                    existingRecordMap[row[0]] = True
+
+            if len(datastring) > 0:
+                keys = datastring.split('\t')
+                def submitkeys(keylist):
+                    if len(keylist) > 0:
+                        sqlstring = "INSERT INTO {0} VALUES ".format(DBTBESC(subsetTable))
+                        sqlstring += ', '.join([valuesString.format(key, subsetid) for key in keylist])
+                        cur.execute(sqlstring)
+                        cur.commit()
+                keysublist = []
+                keyNr = 0
+                for key in keys:
+                    if key not in existingRecordMap:
+                        keysublist.append(key)
+                        if len(keysublist) >= 500:
+                            submitkeys(keysublist)
+                            keysublist = []
+                            calculationObject.SetInfo('Storing', keyNr*1.0/len(keys))
+                        keyNr += 1
+                submitkeys(keysublist)
